Distinguish student files from identical organizer files, consistency in naming convention
